requirements.txt
alembic==1.4.2
appdirs==1.4.4
black==21.4b0
blinker==1.4
cachelib==0.1.1
certifi==2020.12.5
click==7.1.2
cloudinary==1.25.0
dnspython==1.16.0
dominate==2.5.1
email-validator==1.0.5
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-Reuploaded==1.0.0
Flask-Script==2.0.6
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
greenlet==1.0.0
gunicorn==20.1.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.1
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.17.0
python-editor==1.0.4
regex==2021.4.4
six==1.14.0
SQLAlchemy==1.3.16
toml==0.10.2
urllib3==1.26.4
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.0